parser: fix job.Clone()
