+++ parser/model/model_test.go
@@ -18,7 +18,6 @@ import (
 	"fmt"
 	"testing"
 	"time"
-	"unsafe"
 
 	"github.com/pingcap/tidb/parser/charset"
 	"github.com/pingcap/tidb/parser/mysql"
@@ -744,13 +743,4 @@ func TestLocation(t *testing.T) {
 	require.Equal(t, nLoc.String(), "UTC")
 	location := time.FixedZone("UTC", loc1.Offset)
 	require.Equal(t, nLoc, location)
-}
-
-func TestDDLJobSize(t *testing.T) {
-	msg := `Please make sure that the following methods work as expected:
-- SubJob.FromProxyJob()
-- SubJob.ToProxyJob()
-`
-	job := Job{}
-	require.Equal(t, 288, int(unsafe.Sizeof(job)), msg)
 }
